Moved spectrum ui components (#425)
