+++ plugins/ui/src/deephaven/ui/components/__init__.py
@@ -1,26 +1,39 @@
+from .action_button import *
 from .action_group import action_group
 from .action_menu import action_menu
-from .icon import icon
-from .make_component import make_component as component
-from .fragment import fragment
-from .panel import panel
-from .spectrum import *
-from .table import table
-from .dashboard import dashboard
-from .row import row
+from .basic import *
+from .button import *
+from .button_group import *
+from .checkbox import *
 from .column import column
-from .stack import stack
-from .picker import picker
-from .action_group import action_group
-from .section import section
-from .action_menu import action_menu
+from .combo_box import *
+from .content import *
+from .contextual_help import *
+from .dashboard import dashboard
+from .date_picker import *
+from .flex import *
+from .fragment import fragment
+from .icon import icon
 from .item import item
-from .list_view import list_view
+from .item_table_source import item_table_source
 from .list_action_group import list_action_group
 from .list_action_menu import list_action_menu
-from .item_table_source import item_table_source
+from .list_view import list_view
+from .make_component import make_component as component
+from .number_field import *
+from .panel import panel
+from .picker import picker
 from .radio import radio
 from .radio_group import radio_group
+from .range_slider import *
+from .row import row
+from .section import section
+from .slider import *
+from .stack import stack
+from .table import table
+from .text_field import *
+from .toggle_button import *
+from .types import *
 
 from . import html
 
@@ -29,6 +42,7 @@ __all__ = [
     "action_button",
     "action_group",
     "action_menu",
+    "base_element",
     "button",
     "button_group",
     "checkbox",
@@ -60,7 +74,6 @@ __all__ = [
     "row",
     "section",
     "slider",
-    "spectrum_element",
     "stack",
     "switch",
     "table",
