+++ plugins/ui/src/deephaven/ui/components/text_field.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 from typing import Any, Callable
-
-from .events import (
+from .types import (
+    # Accessibility
+    AriaHasPopup,
+    AriaAutoComplete,
+    # Events
     FocusEventCallable,
     KeyboardEventCallable,
-)
-from .layout import (
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
@@ -15,38 +17,36 @@ from .layout import (
     Position,
     LabelPosition,
     Align,
-)
-from .validate import (
+    # Validation
+    TextFieldType,
+    TextFieldInputMode,
     TextFieldValidationState,
     NecessityIndicator,
 )
-from .basic import spectrum_element
-from ...elements import Element
+from .basic import base_element
+from ..elements import Element
 
 
-def number_field(
+def text_field(
+    icon: Element | None = None,
     is_quiet: bool | None = None,
-    hide_stepper: bool | None = None,
-    decrement_aria_label: str | None = None,
-    increment_aria_label: str | None = None,
-    is_wheel_disabled: bool | None = None,
-    # format_options, # omitted because need to connect it to Deephaven formatting options as well
     is_disabled: bool | None = None,
     is_read_only: bool | None = None,
     is_required: bool | None = None,
-    # validation_behaviour, # omitted because validate is not implemented
-    # validate, # omitted because it needs to return a ValidationError synchronously
-    auto_focus: bool | None = None,
-    value: float | None = None,
-    default_value: float | None = None,
-    min_value: float | None = None,
-    max_value: float | None = None,
-    step: float | None = None,
-    label: Any | None = None,
     description: Any | None = None,
     error_message: Any | None = None,
-    validation_state: TextFieldValidationState | None = None,
+    auto_focus: bool | None = None,
+    value: str | None = None,
+    default_value: str | None = None,
+    label: Any | None = None,
+    auto_complete: str | None = None,
+    max_length: int | None = None,
+    min_length: int | None = None,
+    pattern: str | None = None,
+    type: TextFieldType | None = None,
+    input_mode: TextFieldInputMode | None = None,
     name: str | None = None,
+    validation_state: TextFieldValidationState | None = None,
     label_position: LabelPosition = "top",
     label_align: Align = "start",
     necessity_indicator: NecessityIndicator = "icon",
@@ -56,7 +56,7 @@ def number_field(
     on_focus_change: Callable[[bool], None] | None = None,
     on_key_down: KeyboardEventCallable | None = None,
     on_key_up: KeyboardEventCallable | None = None,
-    on_change: Callable[[float], None] | None = None,
+    on_change: Callable[[str], None] | None = None,
     flex: LayoutFlex | None = None,
     flex_grow: Number | None = None,
     flex_shrink: Number | None = None,
@@ -94,37 +94,42 @@ def number_field(
     z_index: Number | None = None,
     is_hidden: bool | None = None,
     id: str | None = None,
+    exclude_from_tab_order: bool | None = None,
+    aria_active_descendant: str | None = None,
+    aria_auto_complete: AriaAutoComplete | None = None,
+    aria_haspopup: AriaHasPopup | None = None,
     aria_label: str | None = None,
     aria_labelledby: str | None = None,
     aria_describedby: str | None = None,
     aria_details: str | None = None,
+    aria_errormessage: str | None = None,
     UNSAFE_class_name: str | None = None,
     UNSAFE_style: CSSProperties | None = None,
     # missing properties that are clipboard or composition events
 ) -> Element:
     """
-    NumberFields allow users to enter a number, and increment or decrement the value using stepper buttons.
+    TextFields are text inputs that allow users to input custom text entries with a keyboard. Various decorations can be displayed around the field to communicate the entry requirements.
 
     Args:
+        icon: An icon to display at the start of the input
         is_quiet: Whether the input should be displayed with a quiet style
-        hide_stepper: Whether to hide the increment and decrement stepper buttons
-        decrement_aria_label: The aria label for the decrement stepper button. If not provided, the default is "Decrement"
-        increment_aria_label: The aria label for the increment stepper button. If not provided, the default is "Increment"
-        is_wheel_disabled: Whether the input should change with scroll
         is_disabled: Whether the input should be disabled
         is_read_only: Whether the input scan be selected but not changed by the user
         is_required: Whether the input is required before form submission
+        description: A description for the field. Provides a hint such as specific requirements for what to choose.
+        error_message: An error message to display when the field is invalid
         auto_focus: Whether the input should be focused on page load
         value: The current value of the input
         default_value: The default value of the input
-        min_value: The minimum value of the input
-        max_value: The maximum value of the input
-        step: The step value for the input
         label: The label for the input
-        description: A description for the field. Provides a hint such as specific requirements for what to choose.
-        error_message: An error message to display when the field is invalid
-        validation_state: Whether the input should display its "valid" or "invalid" state
+        auto_complete: Describes the type of autocomplete functionality the input should provide
+        max_length: The maximum number of characters the input can accept
+        min_length: The minimum number of characters the input can accept
+        pattern: A regular expression that the input's value must match
+        type: The type of input to display
+        input_mode: Hints at the tpye of data that might be entered by the user while editing the element or its contents
         name: The name of the input, used when submitting an HTML form
+        validation_state: Whether the input should display its "valid" or "invalid" state
         label_position: The position of the label relative to the input
         label_align: The alignment of the label relative to the input
         necessity_indicator: Whether the required state should be shown as an icon or text
@@ -172,35 +177,39 @@ def number_field(
         z_index: The stack order of the element.
         is_hidden: Whether the element is hidden.
         id: The unique identifier of the element.
+        exclude_from_tab_order: Whether the element should be excluded from the tab order.
+        aria_active_descendant: Identifies the currently active element when DOM focus is on a composite widget, textbox, group, or application.
+        aria_auto_complete: Indicates whether inputting text could trigger display of one or more predictions of the user's intended value for an input and specifies how predictions would be presented if they are made.
         aria_label: The label for the element.
         aria_labelled_by: The id of the element that labels the current element.
         aria_described_by: The id of the element that describes the current element.
         aria_details: The id of the element that provides additional information about the current element.
+        aria_errormessage: The id of the element that provides an error message for the current element.
         UNSAFE_class_name: A CSS class to apply to the element.
         UNSAFE_style: A CSS style to apply to the element.
     """
 
-    return spectrum_element(
-        "NumberField",
+    return base_element(
+        "TextField",
+        icon=icon,
         is_quiet=is_quiet,
-        hide_stepper=hide_stepper,
-        decrement_aria_label=decrement_aria_label,
-        increment_aria_label=increment_aria_label,
-        is_wheel_disabled=is_wheel_disabled,
         is_disabled=is_disabled,
         is_read_only=is_read_only,
         is_required=is_required,
+        description=description,
+        error_message=error_message,
         auto_focus=auto_focus,
         value=value,
         default_value=default_value,
-        min_value=min_value,
-        max_value=max_value,
-        step=step,
         label=label,
-        description=description,
-        error_message=error_message,
-        validation_state=validation_state,
+        auto_complete=auto_complete,
+        max_length=max_length,
+        min_length=min_length,
+        pattern=pattern,
+        type=type,
+        input_mode=input_mode,
         name=name,
+        validation_state=validation_state,
         label_position=label_position,
         label_align=label_align,
         necessity_indicator=necessity_indicator,
@@ -248,10 +257,15 @@ def number_field(
         z_index=z_index,
         is_hidden=is_hidden,
         id=id,
+        exclude_from_tab_order=exclude_from_tab_order,
+        aria_active_descendant=aria_active_descendant,
+        aria_auto_complete=aria_auto_complete,
+        aria_haspopup=aria_haspopup,
         aria_label=aria_label,
         aria_labelledby=aria_labelledby,
         aria_describedby=aria_describedby,
         aria_details=aria_details,
+        aria_errormessage=aria_errormessage,
         UNSAFE_class_name=UNSAFE_class_name,
         UNSAFE_style=UNSAFE_style,
     )
